Reorganize distributions module (#130)
