+++ numpyro/distributions/__init__.py
@@ -1,41 +1,52 @@
-from numpyro.distributions.constraint_registry import biject_to
 from numpyro.distributions.continuous import (
-    beta,
-    cauchy,
-    expon,
-    gamma,
-    halfcauchy,
-    halfnorm,
-    lognorm,
-    norm,
-    pareto,
-    t,
-    trunccauchy,
-    truncnorm,
-    uniform
+    Beta,
+    Cauchy,
+    Chi2,
+    Dirichlet,
+    Exponential,
+    Gamma,
+    HalfCauchy,
+    LogNormal,
+    Normal,
+    Pareto,
+    StudentT,
+    Uniform
 )
-from numpyro.distributions.discrete import bernoulli, binom, poisson
-from numpyro.distributions.multivariate import categorical, dirichlet, multinomial
+from numpyro.distributions.discrete import (
+    Bernoulli,
+    BernoulliWithLogits,
+    Binomial,
+    BinomialWithLogits,
+    Categorical,
+    CategoricalWithLogits,
+    Multinomial,
+    MultinomialWithLogits,
+    Poisson
+)
+from numpyro.distributions.distribution import Distribution, TransformedDistribution
 
 __all__ = [
-    'beta',
-    'bernoulli',
-    'biject_to',
-    'binom',
-    'cauchy',
-    'categorical',
-    'dirichlet',
-    'expon',
-    'gamma',
-    'halfcauchy',
-    'halfnorm',
-    'lognorm',
-    'multinomial',
-    'norm',
-    'pareto',
-    'poisson',
-    't',
-    'trunccauchy',
-    'truncnorm',
-    'uniform',
+    'Bernoulli',
+    'BernoulliWithLogits',
+    'Beta',
+    'Binomial',
+    'BinomialWithLogits',
+    'Categorical',
+    'CategoricalWithLogits',
+    'Cauchy',
+    'Chi2',
+    'Dirichlet',
+    'Distribution',
+    'Exponential',
+    'Gamma',
+    'HalfCauchy',
+    'LogNormal',
+    'Multinomial',
+    'MultinomialWithLogits',
+    'Normal',
+    'Pareto',
+    'Poisson',
+    'StudentT',
+    'TransformedDistribution',
+    'Uniform',
 ]
