+++ test/test_mcmc.py
@@ -38,12 +38,12 @@ def test_logistic_regression(algo):
     data = random.normal(random.PRNGKey(0), (N, dim))
     true_coefs = np.arange(1., dim + 1.)
     logits = np.sum(true_coefs * data, axis=-1)
-    labels = dist.bernoulli(logits, is_logits=True).rvs(random_state=random.PRNGKey(1))
+    labels = dist.BernoulliWithLogits(logits).sample(random.PRNGKey(1))
 
     def model(labels):
-        coefs = sample('coefs', dist.norm(np.zeros(dim), np.ones(dim)))
+        coefs = sample('coefs', dist.Normal(np.zeros(dim), np.ones(dim)))
         logits = np.sum(coefs * data, axis=-1)
-        return sample('obs', dist.bernoulli(logits, is_logits=True), obs=labels)
+        return sample('obs', dist.BernoulliWithLogits(logits), obs=labels)
 
     init_params, potential_fn, transform_fn = initialize_model(random.PRNGKey(2), model, (labels,), {})
     init_kernel, sample_kernel = hmc(potential_fn, algo=algo)
@@ -62,12 +62,12 @@ def test_beta_bernoulli(algo):
     def model(data):
         alpha = np.array([1.1, 1.1])
         beta = np.array([1.1, 1.1])
-        p_latent = sample('p_latent', dist.beta(alpha, beta))
-        sample('obs', dist.bernoulli(p_latent), obs=data)
+        p_latent = sample('p_latent', dist.Beta(alpha, beta))
+        sample('obs', dist.Bernoulli(p_latent), obs=data)
         return p_latent
 
     true_probs = np.array([0.9, 0.1])
-    data = dist.bernoulli(true_probs).rvs(size=(1000, 2), random_state=random.PRNGKey(1))
+    data = dist.Bernoulli(true_probs).sample(random.PRNGKey(1), size=(1000, 2))
     init_params, potential_fn, transform_fn = initialize_model(random.PRNGKey(2), model, (data,), {})
     init_kernel, sample_kernel = hmc(potential_fn, algo=algo)
     hmc_state = init_kernel(init_params,
@@ -84,12 +84,12 @@ def test_dirichlet_categorical(algo):
 
     def model(data):
         concentration = np.array([1.0, 1.0, 1.0])
-        p_latent = sample('p_latent', dist.dirichlet(alpha=concentration))
-        sample("obs", dist.categorical(p=p_latent), obs=data)
+        p_latent = sample('p_latent', dist.Dirichlet(concentration))
+        sample("obs", dist.Categorical(p_latent), obs=data)
         return p_latent
 
     true_probs = np.array([0.1, 0.6, 0.3])
-    data = dist.categorical(p=true_probs).rvs(size=(2000,), random_state=random.PRNGKey(1))
+    data = dist.Categorical(true_probs).sample(random.PRNGKey(1), size=(2000,))
     init_params, potential_fn, transform_fn = initialize_model(random.PRNGKey(2), model, (data,), {})
     init_kernel, sample_kernel = hmc(potential_fn, algo=algo)
     hmc_state = init_kernel(init_params,
